requirements.txt
fuzzywuzzy==0.14.0
Jinja2==2.8
prettyconf==1.2.3
python-Levenshtein==0.12.0
python-telegram-bot==5.2.0
requests==2.12.1
